+++ platform/nxp/lpc17xx/i2s_base.c
@@ -105,7 +105,7 @@ static void configPins(struct I2sBase *interface,
       config->tx.sck, config->tx.ws, config->tx.sda, config->tx.mclk
   };
 
-  for (unsigned int index = 0; index < ARRAY_SIZE(pinArray); ++index)
+  for (size_t index = 0; index < ARRAY_SIZE(pinArray); ++index)
   {
     if (pinArray[index])
     {
